+++ Sources/VCore/Services and Managers/Keychain Service/KeychainServiceError.swift
@@ -15,7 +15,7 @@ public struct KeychainServiceError: VCoreError, Equatable {
     
     // MARK: VCore Error
     public var code: Int { _code.rawValue }
-    public var description: String { VCoreLocalizationService.shared.localizationProvider.keychainServiceErrorDescription(_code) }
+    public var description: String { VCoreLocalizationManager.shared.localizationProvider.keychainServiceErrorDescription(_code) }
     
     // MARK: Initializers
     init(_ code: Code) {